Configuração JWT para multiplas APIs
