+++ NerdStoreEnterprise.sln
@@ -13,15 +13,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Web", "Web", "{0F4FB6E0-98F
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Identidade", "Identidade", "{BE5FC9F4-0BDA-45D2-84FC-2AFC03075FBB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Identidade.API", "src\services\NSE.Identidade.API\NSE.Identidade.API.csproj", "{C3807557-2E48-4F7F-8CDB-B4CA334699B1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Identidade.API", "src\services\NSE.Identidade.API\NSE.Identidade.API.csproj", "{C3807557-2E48-4F7F-8CDB-B4CA334699B1}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MVC", "MVC", "{1A9A1480-FA65-449C-A133-FE9ED069C3B4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.WebApp.MVC", "src\web\NSE.WebApp.MVC\NSE.WebApp.MVC.csproj", "{D3EE964E-EBF6-4EAD-B808-D16E87F42731}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.WebApp.MVC", "src\web\NSE.WebApp.MVC\NSE.WebApp.MVC.csproj", "{D3EE964E-EBF6-4EAD-B808-D16E87F42731}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{D947836D-C55D-4E12-A8FD-CEC18A18E408}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Core", "src\building blocks\NSE.Core\NSE.Core.csproj", "{654E7DF9-67FB-4160-A0BF-BFA61558C254}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Core", "src\building blocks\NSE.Core\NSE.Core.csproj", "{654E7DF9-67FB-4160-A0BF-BFA61558C254}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Carrinho", "Carrinho", "{F46A6C35-CC85-4595-8A72-C3E03ADE7F45}"
 EndProject
@@ -33,15 +33,19 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Pagamento", "Pagamento", "{
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Pedido", "Pedido", "{A1554471-F624-436F-B06B-77AAA671AC1C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Carrinho.API", "src\services\NSE.Carrinho.API\NSE.Carrinho.API.csproj", "{E8717A25-CEBE-49CE-A88B-63A042D2CCA7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Carrinho.API", "src\services\NSE.Carrinho.API\NSE.Carrinho.API.csproj", "{E8717A25-CEBE-49CE-A88B-63A042D2CCA7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Catalogo.API", "src\services\NSE.Catalogo.API\NSE.Catalogo.API.csproj", "{1EC79BA5-85ED-4E0C-9678-1BDDEC76D5F8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Catalogo.API", "src\services\NSE.Catalogo.API\NSE.Catalogo.API.csproj", "{1EC79BA5-85ED-4E0C-9678-1BDDEC76D5F8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Cliente.API", "src\services\NSE.Cliente.API\NSE.Cliente.API.csproj", "{3A6301D0-2E28-4317-9FE9-B0CAEEB971E4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Cliente.API", "src\services\NSE.Cliente.API\NSE.Cliente.API.csproj", "{3A6301D0-2E28-4317-9FE9-B0CAEEB971E4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Pagamento.API", "src\services\NSE.Pagamento.API\NSE.Pagamento.API.csproj", "{8B3BF348-2083-4432-A71E-AE43FEA903AF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Pagamento.API", "src\services\NSE.Pagamento.API\NSE.Pagamento.API.csproj", "{8B3BF348-2083-4432-A71E-AE43FEA903AF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.Pedido.API", "src\services\NSE.Pedido.API\NSE.Pedido.API.csproj", "{2A650EA4-081A-4DD1-A3BB-278185A6DE70}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NSE.Pedido.API", "src\services\NSE.Pedido.API\NSE.Pedido.API.csproj", "{2A650EA4-081A-4DD1-A3BB-278185A6DE70}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{27B54CA8-2E57-4E30-9D07-5F7DABC7B5E4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NSE.WebApi.Core", "src\building blocks\NSE.WebApi.Core\NSE.WebApi.Core.csproj", "{8ED28603-AA13-4341-A296-10DD8C0AC07F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -81,6 +85,10 @@ Global
 		{2A650EA4-081A-4DD1-A3BB-278185A6DE70}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2A650EA4-081A-4DD1-A3BB-278185A6DE70}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2A650EA4-081A-4DD1-A3BB-278185A6DE70}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8ED28603-AA13-4341-A296-10DD8C0AC07F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8ED28603-AA13-4341-A296-10DD8C0AC07F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8ED28603-AA13-4341-A296-10DD8C0AC07F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8ED28603-AA13-4341-A296-10DD8C0AC07F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -105,6 +113,8 @@ Global
 		{3A6301D0-2E28-4317-9FE9-B0CAEEB971E4} = {20515D8F-E606-4D81-A7BA-6D2480F845BA}
 		{8B3BF348-2083-4432-A71E-AE43FEA903AF} = {274FCC3A-927F-4BF0-8E21-16108BC5C6D8}
 		{2A650EA4-081A-4DD1-A3BB-278185A6DE70} = {A1554471-F624-436F-B06B-77AAA671AC1C}
+		{27B54CA8-2E57-4E30-9D07-5F7DABC7B5E4} = {AFB42F33-7971-4D2B-A2E7-C898052FEEB2}
+		{8ED28603-AA13-4341-A296-10DD8C0AC07F} = {27B54CA8-2E57-4E30-9D07-5F7DABC7B5E4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F903A11D-C363-45D0-A88E-10A0ED50919E}
