+++ src/building blocks/NSE.WebApi.Core/Identidade/AppSettings.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace NSE.Identidade.API.Extensions
+﻿namespace NSE.WebApi.Core.Identidade
 {
     public class AppSettings
     {
