+++ src/services/NSE.Catalogo.API/Configuration/ApiConfig.cs
@@ -5,6 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NSE.Catalogo.API.Data;
+using NSE.WebApi.Core.Identidade;
 
 namespace NSE.Catalogo.API.Configuration
 {
@@ -34,6 +35,9 @@ namespace NSE.Catalogo.API.Configuration
             app.UseHttpsRedirection();
 
             app.UseRouting();
+
+            app.UseAuthConfiguration();
+            app.UseCors("Total");
 
             app.UseAuthorization();
