+++ src/services/NSE.Catalogo.API/appsettings.Development.json
@@ -8,5 +8,11 @@
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     }
+  },
+  "AppSettings": {
+    "Secret": "MEUSEGREDOSUPERSECRETO",
+    "expiracaoHoras": 2,
+    "Emissor": "MeuSistema",
+    "ValidoEm": "https://localhost"
   }
 }
